testing exit script

+++ automate-scan.sh
@@ -18,25 +18,39 @@ function pingsweep() {
 # Calling the pingsweep function and run it inside a for loop to each target=host in the loop.
 
 echo "Starting pingsweep"
+
 for hosts in $(cat list.txt); do
     pingsweep $hosts
 done
+
 echo "Finishing pingsweep"
 
 echo "List of reachable hosts"
+
 host_reachable=$(cat /tmp/reachable | awk '{print $1}')
+
 echo $host_reachable | tr ' ' ' \n' | sort
 
-echo "Starting network scan"
+status_host_reachable=$?
 
-# Nmap scanning inside a loop to get all host reachables after the pingsweep (I don't care the unrechable targets at this moment)
+    if [ $status_host_reachable -eq 0 ]; then
 
-for scan in $host_reachable; do
-    nmap -sV --script=banner $scan
-done
+        echo "Starting network scan"
 
-# Deleting file from tmp directory
+        # Nmap scanning inside a loop to get all host reachables after the pingsweep (I don't care the unrechable targets at this moment)
 
-rm -f /tmp/reachable
-echo "Finishing network scan"
+        for scan in $host_reachable; do
+            nmap -sV --script=banner $scan
+        done
+
+        # Deleting file from tmp directory
+
+        rm -f /tmp/reachable
+        echo "Finishing network scan"
+
+    else
+
+        echo "Error"
+        exit 1
+    fi
